Renamed CDAMatchineEngine; added a TwoSidedAuction trait.
